requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
Requests==2.31.0
WTForms==3.0.1

python-dotenv~=1.0.0